requirements.txt
langchain==0.1.16
langchain-google-genai==0.0.11
faiss-cpu
python-dotenv